fastapi_backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
blis==1.3.3
catalogue==2.0.10
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
cymem==2.0.13
de_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.8.0/de_core_news_sm-3.8.0-py3-none-any.whl#sha256=fec69fec52b1780f2d269d5af7582a5e28028738bd3190532459aeb473bfa3e7
fastapi==0.129.0
filelock==3.24.0
fsspec==2026.2.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.15
networkx==3.6.1
numpy==2.4.2
packaging==26.0
pillow==12.1.1
preshed==3.0.12
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
requests==2.32.5
rich==14.3.2
sentencepiece==0.2.1
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
starlette==0.52.1
sympy==1.14.0
thinc==8.3.10
torch==2.10.0
torchtext==0.6.0
torchvision==0.25.0
tqdm==4.67.3
typer==0.23.1
typer-slim==0.23.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
wasabi==1.1.3
weasel==0.4.3
wrapt==2.1.1
